+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [NEXT]
 * BREAKING CHANGES: change `forceTileSize` type double to Size.
+* Add support to offsetX and offsetY in TileMap layers.
 
 ## [0.6.17]
 * hotfix Tiled - Support multiTileset.